Login clean up
